+++ src/routing/pages/viewer/Viewer.vue
@@ -4,10 +4,10 @@
         <ViewerRecipes v-else-if="$viewer.list" />
 
         <transition
-            :enter-active-class="starting ? '' : 'transition duration-1000'"
+            :enter-active-class="!$ui.animations || !starting ? 'transition duration-1000' : ''"
             enter-from-class="opacity-0"
             enter-to-class="opacity-100"
-            :leave-active-class="starting ? '' : 'transition duration-1000'"
+            :leave-active-class="!$ui.animations || !starting ? 'transition duration-1000' : ''"
             leave-from-class="opacity-100"
             leave-to-class="opacity-0"
         >