+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
+- Settings modal.
 - [ISO 8601](https://en.wikipedia.org/wiki/ISO_8601#Durations) durations support.
 
 ### Fixed